Sichi fixes, increased spawn rate of flowers, added lanterns to crypt slayer.
Added Lanterns to Relearn Crafts incase people are passed the point to gain the lantern.
Throw Script fix, can't throw through walls.
New DarknessMap Script to apply to maps that are in caves/dungeons.
GodMode now turns off darkness.

+++ Chaos.Common/Definitions/Enums.cs
@@ -1244,7 +1244,7 @@ public enum LightLevel : byte
 }
 
 [Flags]
-public enum MapFlags : ulong
+public enum MapFlags : byte
 {
     None = 0,
     Snow = 1,
